requirements.txt
pillow
langchain
langchain-anthropic
langchain-classic
pytesseract
ipython
python-dotenv
